requirements/base.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
django-phonenumber-field[phonenumbers]==5.0.0
djangorestframework==3.12.1
Markdown==3.3.1
django-filter==2.4.0